Fix home grafana dashboard for MCOA (#2033)
* WIP: Dynamically set grafana dash to correct one

Ideally we'd want the Grafana home dashboard to be set dynamically based
on whehther we are running "normal" mode or MCOA.

This commit is an attempt to do just that, however it currently doesn't
work well due to the fact that kubernetes configmaps are not "updated"
on label/annotation changes, which causes the grafana-dashboard-loader
to not correctly pick up the changes.

The implementation currently uses labels, but we should probably use
annotations instead.

Signed-off-by: Jacob Baungard Hansen <jacobbaungard@redhat.com>

* update cm on labals change, fixes

Signed-off-by: Thibault Mange <22740367+thibaultmg@users.noreply.github.com>

* use annotations for set-home-dashboard

Signed-off-by: Thibault Mange <22740367+thibaultmg@users.noreply.github.com>

---------

Signed-off-by: Jacob Baungard Hansen <jacobbaungard@redhat.com>
Signed-off-by: Thibault Mange <22740367+thibaultmg@users.noreply.github.com>
Co-authored-by: Jacob Baungard Hansen <jacobbaungard@redhat.com>